renal_framework/requirements.txt
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asttokens==3.0.1
attrs==25.4.0
blis==1.3.3
catalogue==2.0.10
catboost==1.2.8
certifi==2026.1.4
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
cloudpickle==3.1.2
comm==0.2.3
confection==0.1.5
contourpy==1.3.3
cramjam==2.11.0
cycler==0.12.1
cymem==2.0.13
debugpy==1.8.20
decorator==5.2.1
distlib==0.4.0
dotenv==0.9.9
drift-resilient-tabpfn @ git+https://github.com/automl/Drift-Resilient_TabPFN.git@a6e75afb82d13e7abb46deade3669c4106b3d636
einops==0.8.2
executing==2.2.1
faiss-cpu==1.13.2
fastparquet==2025.12.0
filelock==3.20.3
folktables==0.0.12
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.1.0
future==1.0.0
gensim==4.4.0
graphviz==0.21
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.0
hyperopt==0.2.7
identify==2.6.16
idna==3.11
imutils==0.5.4
iniconfig==2.3.0
ipykernel==7.1.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
kditransform==1.0.0
kiwisolver==1.4.9
liac-arff==2.5.0
lightgbm==3.3.5
lightning-utilities==0.15.2
llvmlite==0.46.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
minio==7.2.20
mpmath==1.3.0
msgpack==1.1.2
murmurhash==1.0.15
narwhals==2.16.0
nest-asyncio==1.6.0
networkx==3.6.1
nodeenv==1.10.0
numba==0.63.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
opencv-contrib-python==4.13.0.90
openml==0.14.1
packaging==26.0
pacmap==0.9.1
pandas==3.0.0
parso==0.8.5
pexpect==4.9.0
pillow==12.1.0
platformdirs==4.5.1
plotly==6.5.2
pluggy==1.6.0
pre_commit==4.5.1
preshed==3.0.12
prompt_toolkit==3.0.52
protobuf==6.33.5
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
py4j==0.10.9.9
pyarrow==23.0.0
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytorch-widedeep==1.3.2
PyYAML==6.0.3
pyzmq==27.1.0
ray==2.6.3
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.5.2
scipy==1.17.0
seaborn==0.13.2
shap==0.50.0
shellingham==1.5.4
six==1.17.0
slicer==0.0.8
smart_open==7.5.0
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2
stack-data==0.6.3
sympy==1.14.0
tabulate==0.9.0
thinc==8.3.10
threadpoolctl==3.6.0
tigramite==5.2.10.1
tokenizers==0.22.2
torch==2.1.2
torchmetrics==1.2.0
torchvision==0.16.2
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
transformers==5.0.0
triton==2.1.0
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
virtualenv==20.36.1
wasabi==1.1.3
wcwidth==0.5.3
weasel==0.4.3
wrapt==2.1.1
xgboost==3.1.3
xmltodict==1.0.2